More test projects

+++ OpenCLTest2/src/main/java/no/plasmid/opencl/PerlinNoise.java
@@ -2,12 +2,12 @@ package no.plasmid.opencl;
 
 public class PerlinNoise {
 
-	double persistence, frequency, amplitude;
+	float persistence, frequency, amplitude;
 	int octaves, randomseed;
 	  
 	public PerlinNoise()
 	{
-		this(0.0, 0.0, 0.0, 0, 0);
+		this(0.0f, 0.0f, 0.0f, 0, 0);
 	}
 
 	/**
@@ -18,7 +18,7 @@ public class PerlinNoise {
 	 * @param octaves 
 	 * @param randomseed random seed
 	 */
-	public PerlinNoise(double persistence, double frequency, double amplitude, int octaves, int randomseed)
+	public PerlinNoise(float persistence, float frequency, float amplitude, int octaves, int randomseed)
 	{
 		this.persistence = persistence;
 		this.frequency = frequency;
@@ -27,16 +27,16 @@ public class PerlinNoise {
 		this.randomseed = randomseed;
 	}
 
-	public double getHeight(double x, double y) {
+	public float getHeight(float x, float y) {
 		return amplitude * total(x, y);
 	}
 	
-	private double total(double x, double y)
+	private float total(float x, float y)
 	{
 		//properties of one octave (changing each loop)
-		double t = 0.0f;
-		double amplitude = 1;
-		double freq = frequency;
+		float t = 0.0f;
+		float amplitude = 1;
+		float freq = frequency;
 
 		for(int k = 0; k < octaves; k++)  {
 			t += genValue(y * freq + randomseed, x * freq + randomseed) * amplitude;
@@ -47,62 +47,62 @@ public class PerlinNoise {
 		return t;
 	}
 	
-	private double genValue(double x, double y) {
+	private float genValue(float x, float y) {
 		int Xint = (int)x;
 		int Yint = (int)y;
-		double Xfrac = x - Xint;
-		double Yfrac = y - Yint;
+		float Xfrac = x - Xint;
+		float Yfrac = y - Yint;
 
 		//noise values
-		double n01 = noise(Xint-1, Yint-1);
-		double n02 = noise(Xint+1, Yint-1);
-		double n03 = noise(Xint-1, Yint+1);
-		double n04 = noise(Xint+1, Yint+1);
-		double n05 = noise(Xint-1, Yint);
-		double n06 = noise(Xint+1, Yint);
-		double n07 = noise(Xint, Yint-1);
-		double n08 = noise(Xint, Yint+1);
-		double n09 = noise(Xint, Yint);
+		float n01 = noise(Xint-1, Yint-1);
+		float n02 = noise(Xint+1, Yint-1);
+		float n03 = noise(Xint-1, Yint+1);
+		float n04 = noise(Xint+1, Yint+1);
+		float n05 = noise(Xint-1, Yint);
+		float n06 = noise(Xint+1, Yint);
+		float n07 = noise(Xint, Yint-1);
+		float n08 = noise(Xint, Yint+1);
+		float n09 = noise(Xint, Yint);
 
-		double n12 = noise(Xint+2, Yint-1);
-		double n14 = noise(Xint+2, Yint+1);
-		double n16 = noise(Xint+2, Yint);
+		float n12 = noise(Xint+2, Yint-1);
+		float n14 = noise(Xint+2, Yint+1);
+		float n16 = noise(Xint+2, Yint);
 
-		double n23 = noise(Xint-1, Yint+2);
-		double n24 = noise(Xint+1, Yint+2);
-		double n28 = noise(Xint, Yint+2);
+		float n23 = noise(Xint-1, Yint+2);
+		float n24 = noise(Xint+1, Yint+2);
+		float n28 = noise(Xint, Yint+2);
 
-		double n34 = noise(Xint+2, Yint+2);
+		float n34 = noise(Xint+2, Yint+2);
 
 		//find the noise values of the four corners
-		double x0y0 = 0.0625*(n01+n02+n03+n04) + 0.125*(n05+n06+n07+n08) + 0.25*(n09);  
-		double x1y0 = 0.0625*(n07+n12+n08+n14) + 0.125*(n09+n16+n02+n04) + 0.25*(n06);  
-		double x0y1 = 0.0625*(n05+n06+n23+n24) + 0.125*(n03+n04+n09+n28) + 0.25*(n08);  
-		double x1y1 = 0.0625*(n09+n16+n28+n34) + 0.125*(n08+n14+n06+n24) + 0.25*(n04);  
+		float x0y0 = 0.0625f*(n01+n02+n03+n04) + 0.125f*(n05+n06+n07+n08) + 0.25f*(n09);  
+		float x1y0 = 0.0625f*(n07+n12+n08+n14) + 0.125f*(n09+n16+n02+n04) + 0.25f*(n06);  
+		float x0y1 = 0.0625f*(n05+n06+n23+n24) + 0.125f*(n03+n04+n09+n28) + 0.25f*(n08);  
+		float x1y1 = 0.0625f*(n09+n16+n28+n34) + 0.125f*(n08+n14+n06+n24) + 0.25f*(n04);  
 
 		//interpolate between those values according to the x and y fractions
-		double v1 = interpolate(x0y0, x1y0, Xfrac); //interpolate in x direction (y)
-		double v2 = interpolate(x0y1, x1y1, Xfrac); //interpolate in x direction (y+1)
-		double fin = interpolate(v1, v2, Yfrac);  //interpolate in y direction
+		float v1 = interpolate(x0y0, x1y0, Xfrac); //interpolate in x direction (y)
+		float v2 = interpolate(x0y1, x1y1, Xfrac); //interpolate in x direction (y+1)
+		float fin = interpolate(v1, v2, Yfrac);  //interpolate in y direction
 
 		return fin;
 	}
 
-	private double interpolate(double x, double y, double a) {
-		double negA = 1.0 - a;
-		double negASqr = negA * negA;
-		double fac1 = 3.0 * (negASqr) - 2.0 * (negASqr * negA);
-		double aSqr = a * a;
-		double fac2 = 3.0 * aSqr - 2.0 * (aSqr * a);
+	private float interpolate(float x, float y, float a) {
+		float negA = 1.0f - a;
+		float negASqr = negA * negA;
+		float fac1 = 3.0f * (negASqr) - 2.0f * (negASqr * negA);
+		float aSqr = a * a;
+		float fac2 = 3.0f * aSqr - 2.0f * (aSqr * a);
 
 		return x * fac1 + y * fac2; //add the weighted factors
 	}
 
-	private double noise(int x, int y) {
+	private float noise(int x, int y) {
 		int n = x + y * 57;
 		n = (n << 13) ^ n;
 		int t = (n * (n * n * 15731 + 789221) + 1376312589) & 0x7fffffff;
-		return 1.0 - (double)t * 0.931322574615478515625e-9;/// 1073741824.0);
+		return (float)(1.0 - t * 0.931322574615478515625e-9);/// 1073741824.0);
 	}	
 	
 }
